perf: keep source_ids uint8 through the input path; raise worker default to 32
The species data is 0.7% dense (mean 1.4 presences out of 200), but source_ids
was collated as int64 — 26 MB per batch of ~99% zeros, serialised through worker
IPC and copied to the device every step. Collate it as uint8 (3.3 MB) and widen
it on-device, where the flat (state, species) gather index is now built once per
forward instead of once per layer.

Measured on H100 (10M-row shard, batch 256, 32 workers, compile, same GPU):
7420 -> 7616 samples/s (+2.6%), loss/accuracy identical at every logged step.
Modest, because the batch bytes were not the dominant input cost — but it is
free and cuts memory traffic.

Also raise the --num_workers default from min(8, cpu_count) to
min(32, cpu_count): after the source K/V gather made the step ~2x cheaper the
input pipeline became the constraint again, and 16->32 workers is +17%
(6333 -> 7421 samples/s), with 48 and 64 flat. The earlier "plateaus at 8"
measurement was taken when the GPU was the bottleneck.

torch.sparse is deliberately not used: at these shapes its kernels are slower
than dense ones. The sparsity is exploited structurally instead — narrow dtypes
here, and the 3*S-distinct-rows gather in the previous commit.

+++ stemlm/data.py
@@ -571,7 +571,10 @@ class JSDMDataCollator:
             )
         target_ids[indices_replaced] = 2
 
-        source_ids = source_species.long()
+        # 0/1 presence: keep it uint8 through worker IPC and the host copy
+        # (int64 would be 8x the bytes for a tensor that is ~99% zeros);
+        # the model widens it on-device where the gather needs an index.
+        source_ids = source_species.to(torch.uint8)
 
         labels[~masked_indices] = -100
 
@@ -623,7 +626,10 @@ class AbsenceMaskCollator(JSDMDataCollator):
         target_ids = target_species.long()
         target_ids[masked] = 2
 
-        source_ids = source_species.long()
+        # 0/1 presence: keep it uint8 through worker IPC and the host copy
+        # (int64 would be 8x the bytes for a tensor that is ~99% zeros);
+        # the model widens it on-device where the gather needs an index.
+        source_ids = source_species.to(torch.uint8)
 
         labels = target_species.clone()
         labels[~masked] = -100
@@ -668,7 +674,10 @@ class FixedPValCollator(JSDMDataCollator):
         target_ids = target_species.long()
         target_ids[masked] = 2
 
-        source_ids = source_species.long()
+        # 0/1 presence: keep it uint8 through worker IPC and the host copy
+        # (int64 would be 8x the bytes for a tensor that is ~99% zeros);
+        # the model widens it on-device where the gather needs an index.
+        source_ids = source_species.to(torch.uint8)
 
         labels = target_species.clone()
         labels[~masked] = -100
